backend/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
audioread==3.0.1
av==12.3.0
bcrypt==5.0.0
bleach==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
ctranslate2==4.6.0
cycler==0.12.1
datasets==2.14.6
decorator==5.2.1
dill==0.3.7
diskcache==5.6.3
fastapi==0.104.1
faster-whisper==1.0.3
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2023.10.0
greenlet==3.2.4
h11==0.16.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.10.1
llama-cpp-python==0.3.16
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.15
networkx==3.5
noisereduce==3.0.2
numba==0.62.1
numpy==2.3.3
onnxruntime==1.23.1
packaging==25.0
pandas==2.1.3
peft==0.6.2
pillow==11.3.0
pip==25.2
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.9
pyarrow==21.0.0
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
RestrictedPython==8.0
safetensors==0.6.2
scikit-learn==1.3.2
scipy==1.16.2
setuptools==80.9.0
silero-vad==6.0.0
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==1.0.0
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.41.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
webencodings==0.5.1
xxhash==3.6.0
yarl==1.22.0